NLP, LLM and GenAI/LlamaIndex/01/requirements.txt
llama-index
python-dotenv
html2text
llama-index-llms-openai